requirements.txt
robotframework>=4.0.3
python-dotenv>=1.0.0
pyodbc>=4.0.32